requirements.txt
opencv-python
numpy
matplotlib
tk
pillow
